Randomize test node order/source
The nature of the random walk meant that
choosing the last node in the `able` set
was a really good strategy, and the programs
converged to that in a single generation.

With the random changes, I got a apparently
better program in 13 generations, though it only
ran against 9 graphs. Need to see how to read
push programs from a file so I can retest it.